+++ CHANGES.md
@@ -8,6 +8,8 @@ the one you find pypi, but its the one in development.
 =====
 * Modify README to show the right systemd min version.
 * Added `pystemd.journal` with methods `log` and `sendv`.
+* Added `booted` support in `pystemd.daemon` to know if a system was booted with
+  systemd.
 
 0.6.0
 =====